Develop generic nested container range, use for strategies (#683)
Extends the nested range developed for information sets to a generic template, and uses it also to iterate the strategies of a game.

+++ src/core/function.cc
@@ -43,7 +43,7 @@ using namespace Gambit;
 // vector perpendicular to the plane, then subtracting to compute the
 // component parallel to the plane.)
 //
-void FunctionOnSimplices::Project(Vector<double> &x, const Array<int> &lengths) const
+void FunctionOnSimplices::Project(Vector<double> &x, const Array<size_t> &lengths) const
 {
   int index = 1;
   for (size_t part = 1; part <= lengths.size(); part++) {
